% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{A320data}
\alias{A320data}
\title{QAR data from A320}
\format{
A data table with 1218 rows and 103 variables:
}
\usage{
A320data
}
\description{
QAR data from Airbus 320 flights that contains the maximum vertical acceleration during landing as the response and other measurements as covariates. Detailed explanation of feature code are in the supplementary material of the paper.
}
\keyword{datasets}
