% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{A380data}
\alias{A380data}
\title{QAR data from A380}
\format{
A data table with 1218 rows and 103 variables:
}
\usage{
A380data
}
\description{
QAR data from Airbus 320 flights contains the maximum vertical acceleration during landing as the response and other measurements as covariates. A detailed explanation of the feature codes is in the supplementary material of the paper.
}
\keyword{datasets}
