% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cdloss}
\alias{cdloss}
\title{Quantile loss}
\usage{
cdloss(x, y, beta, u)
}
\arguments{
\item{x}{Input matrix. The convariates of samples.}

\item{y}{Input vector. The responses of samples.}

\item{beta}{Input vector. The coefficient.}

\item{u}{Input scalar. The quantile level.}
}
\value{
A scalar. The quantile loss based on the given inputs.
}
\description{
Calculate the quantile loss based on the given samples, coefficient, and quantile level.
}
