% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ChooseBandwidth}
\alias{ChooseBandwidth}
\title{Select bandwidth h by cross validation}
\usage{
ChooseBandwidth(x, y, u, control = list())
}
\arguments{
\item{x}{design matrix which is in size of n*p.}

\item{y}{response vector.}

\item{u}{A scalar for the quantile level.}

\item{control}{The list object for control, see \code{tlqr.control}.}
}
\description{
Select bandwidth h by cross validation
}
