% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{DetectPoolQR}
\alias{DetectPoolQR}
\title{Main function for informative sources detection among all dataset for
Trans-pooling QR}
\usage{
DetectPoolQR(
  x.target,
  y.target,
  x.source,
  y.source,
  u,
  epsilon = 0.01,
  ninfo = NULL,
  control = list()
)
}
\arguments{
\item{x.target}{A n*p matrix as the covariate from the target population.}

\item{y.target}{A vector with length n as the response from the target
population.}

\item{x.source}{A list object that each element is a n*p matrix as the
covariate from informative sources.}

\item{y.source}{A list object that each element is a vector with length n
as the covariate from informative sources.}

\item{u}{A scalar in (0,1). The quantile level for all models.}

\item{epsilon}{A scalar. Default is 0.01. The strict level for informative
sources detectiond, larger the value is, less strict the procedure is.}

\item{ninfo}{An integar. Default is NULL. The given number of informative
sources under pseduo running.}

\item{control}{The list object for control, see \code{tlqr.control}.}
}
\value{
A vector of index. The index of the informative sources.
}
\description{
Main function for informative sources detection among all dataset for
Trans-pooling QR
}
