% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{DetectTransferQR}
\alias{DetectTransferQR}
\title{Trans-Lasso QR with informative set detection}
\usage{
DetectTransferQR(
  x.target,
  y.target,
  x.source,
  y.source,
  u.target,
  u.source,
  epsilon = 0.01,
  ninfo = NULL,
  lambda.det = NULL,
  lambda1 = NULL,
  lambda2 = NULL,
  control = list()
)
}
\arguments{
\item{x.target}{A n*p matrix as the covariate from the target population.}

\item{y.target}{A vector with length n as the response from the target
population.}

\item{x.source}{A list object that each element is a n*p matrix as the
covariate from informative sources.}

\item{y.source}{A list object that each element is a vector with length n
as the covariate from informative sources.}

\item{u.target}{A scalar in (0,1). The quantile level for model of target
population.}

\item{u.source}{A vector that contains the quantile level for informative
populations.}

\item{epsilon}{A scalar. Default is 0.01. The strict level for informative
sources detectiond, larger the value is, less strict the procedure is.}

\item{ninfo}{An integar. Default is NULL. The given number of informative
sources under pseduo running.}

\item{lambda.det}{A scalar for the penalty strength in detection procedure.
Default is null, if null, a cross validation will be carried for searching.}

\item{lambda1}{A scalar for the penalty strength in fusion step. Default
is null, if null, a cross validation will be carried for searching.}

\item{lambda2}{A scalar for the penalty strength in debiasing step. Default
is null, if null, lambda2 will be calculated based on lambda1 according to
the proportion of sample sizes as described in our paper.}

\item{control}{The list object for control, see \code{tlqr.control}.}
}
\value{
A vector. The corresponding transfer learning estimator based on
the given auxiliary sources.
}
\description{
Trans-Lasso QR with informative set detection
}
