% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{FusionQR}
\alias{FusionQR}
\title{Build fusion model between datasets without debias procedure}
\usage{
FusionQR(
  x.target,
  y.target,
  x.source,
  y.source,
  u.target,
  u.source,
  lambda1 = NULL,
  control = list()
)
}
\arguments{
\item{x.target}{A n*p matrix as the covariate from the target population.}

\item{y.target}{A vector with length n as the response from the target
population.}

\item{x.source}{A list object that each element is a n*p matrix as the
covariate from informative sources.}

\item{y.source}{A list object that each element is a vector with length n
as the covariate from informative sources.}

\item{u.target}{A scalar in (0,1). The quantile level for model of target
population.}

\item{u.source}{A vector that contains the quantile level for informative
populations.}

\item{lambda1}{A scalar for the penalty strength in fusion step. Default
is null, if null, a cross validation will be carried for searching.}

\item{control}{The list object for control, see \code{tlqr.control}.}
}
\value{
A vector. The fusion estimator withou debias procedure based on the
given auxiliary sources.
}
\description{
Build fusion model between datasets without debias procedure
}
