% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{hchoose}
\alias{hchoose}
\title{Select bandwidth h by cross validation}
\usage{
hchoose(x, y, u)
}
\arguments{
\item{x}{design matrix which is in size of n*p.}

\item{y}{response vector.}

\item{u}{A scalar for the quantile level.}
}
\description{
Select bandwidth h by cross validation
}
