% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{info_detect.pool}
\alias{info_detect.pool}
\title{Main function for informative sources detection among all dataset for
Trans-pooling QR}
\usage{
info_detect.pool(
  x_target,
  y_target,
  x_aux_bd,
  y_aux_bd,
  u,
  epsilon = 0.01,
  parallel = TRUE,
  ncore = 10,
  info_num = NULL,
  seed = 111,
  verbose = FALSE
)
}
\arguments{
\item{x_target}{A n*p matrix as the covariate from the target population.}

\item{y_target}{A vector with length n as the response from the target
population.}

\item{x_aux_bd}{A list object that each element is a n*p matrix as the
covariate from informative sources.}

\item{y_aux_bd}{A list object that each element is a vector with length n
as the covariate from informative sources.}

\item{u}{A scalar in (0,1). The quantile level for all models.}

\item{epsilon}{A scalar. Default is 0.01. The strict level for informative
sources detectiond, larger the value is, less strict the procedure is.}

\item{parallel}{A logic variable, default is TRUE}

\item{ncore}{The integer, the number of cores used for parallel computation.}

\item{info_num}{An integar. Default is NULL. The given number of informative
sources under pseduo running.}

\item{seed}{A integer variable for the random seed, default is 111.}

\item{verbose}{A logic variable, default is FALSE.}
}
\value{
A vector of index. The index of the informative sources.
}
\description{
Main function for informative sources detection among all dataset for
Trans-pooling QR
}
