% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{info_detect}
\alias{info_detect}
\title{Main function for informative sources detection among all dataset}
\usage{
info_detect(
  x_target,
  y_target,
  x_aux_bd,
  y_aux_bd,
  u_target,
  u_aux_bd,
  epsilon = 0.01,
  mode = "real",
  psd = FALSE,
  info_num = 2
)
}
\arguments{
\item{x_target}{A n*p matrix as the covariate from the target population.}

\item{y_target}{A vector with length n as the response from the target population.}

\item{x_aux_bd}{A list object that each element is a n*p matrix as the covariate from informative sources.}

\item{y_aux_bd}{A list object that each element is a vector with length n as the covariate from informative sources.}

\item{u_target}{A scalar. The quantile level for model of target population.}

\item{u_aux_bd}{A vector that contains the quantile level for informative populations.}

\item{epsilon}{A scalar. Default is 0.01. The strict level for informative sources detectiond, larger the value is, less strict the procedure is.}

\item{mode}{A string. Default is 'real'. A model controler to tell 'simulation' vs 'real', this lead to different operations for tunning parameters of single-source modeling.}

\item{psd}{A boolean variable. Whether the procedure is pseudo.}

\item{info_num}{A integar. The given number of informative sources under pseduo running.}
}
\value{
A vector of index. The index of the informative sources.
}
\description{
Main function for informative sources detection among all dataset
}
