% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{K}
\alias{K}
\title{Build the smoothing kernel for estimation of density}
\usage{
K(x)
}
\arguments{
\item{x}{Numeric variable}
}
\value{
Numeric scalar
}
\description{
Build the smoothing kernel for estimation of density
}
