% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{QuantileLoss}
\alias{QuantileLoss}
\title{Calculate the quantile loss based on the given samples, coefficient, and the
quantile level}
\usage{
QuantileLoss(x, y, beta, u)
}
\arguments{
\item{x}{Input design matrix. The convariates of samples.}

\item{y}{Input response vector. The responses of samples.}

\item{beta}{Input vector. The coefficient.}

\item{u}{Input scalar. The quantile level.}
}
\value{
A scalar. The quantile loss based on the given inputs.
}
\description{
Calculate the quantile loss based on the given samples, coefficient, and the
quantile level
}
