% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rq.fusion.pool}
\alias{rq.fusion.pool}
\title{Build fusion model for Trans-pooling QR between datasets without debias
procedure}
\usage{
rq.fusion.pool(x_target, y_target, x_aux_bd, y_aux_bd, u)
}
\arguments{
\item{x_target}{A n*p matrix as the covariate from the target population.}

\item{y_target}{A vector with length n as the response from the target
population.}

\item{x_aux_bd}{A list object that each element is a n*p matrix as the
covariate from informative sources.}

\item{y_aux_bd}{A list object that each element is a vector with length n
as the covariate from informative sources.}

\item{u}{A scalar in (0,1). The quantile level for all models.}
}
\value{
A vector. The fusion estimator withou debias procedure based on the
given auxiliary sources.
}
\description{
Build fusion model for Trans-pooling QR between datasets without debias
procedure
}
