% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rq.transfer}
\alias{rq.transfer}
\title{Build the function for Oracle Trans-Lasso QR with a given informative sources set}
\usage{
rq.transfer(
  x_target,
  y_target,
  x_aux_bd,
  y_aux_bd,
  u_target,
  u_aux_bd,
  mode = "real"
)
}
\arguments{
\item{x_target}{A n*p matrix as the covariate from the target population.}

\item{y_target}{A vector with length n as the response from the target population.}

\item{x_aux_bd}{A list object that each element is a n*p matrix as the covariate from informative sources.}

\item{y_aux_bd}{A list object that each element is a vector with length n as the covariate from informative sources.}

\item{u_target}{A scalar. The quantile level for model of target population.}

\item{u_aux_bd}{A vector that contains the quantile level for informative populations.}

\item{mode}{A string. Default is 'real'. A model controler to tell 'simulation' vs 'real', this lead to different operations for tunning parameters of single-source modeling.}
}
\value{
A vector. The corresponding transfer learning estimator based on the given auxiliary sources.
}
\description{
Build the function for Oracle Trans-Lasso QR with a given informative sources set
}
