% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{smoothy}
\alias{smoothy}
\title{Get the smoothed response}
\usage{
smoothy(x, y, u, maxit = 100, reltol = 0.05)
}
\arguments{
\item{x}{A n*p matrix as the covariate for response smoothing.}

\item{y}{A vector with length n as the response that needs to be smoothed.}

\item{u}{A scalar. The quantile level.}

\item{maxit}{A integer. The maximum iteration number.}

\item{reltol}{A scalar. Default is 1e-2. The relative tolerance for the stopping criterion.}
}
\value{
y: A vector as the smoothed response. h: A scalar as the selected bandwidth. converge: The convergence status, 0 is succuessful convergence.
}
\description{
Get the smoothed response (from the quantile model to least square model) variable by iterations.
}
