% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tlqr.control}
\alias{tlqr.control}
\title{Internal tlqr parameters}
\usage{
tlqr.control(
  h = NULL,
  hic = NULL,
  detect.parallel = TRUE,
  compute.parallel = FALSE,
  ncore = 20,
  seed = 111,
  lowc = 0.05,
  upc = 1,
  thresh = 1e-10,
  length.out = 25,
  max.it = 1e+06,
  max.round = 5,
  verbose = FALSE
)
}
\arguments{
\item{h}{A vector with length (K+1), where K is the number of auxiliary
datasets. It pre-determine the bandwidth. The default is \code{NULL},
if \code{NULL}, cross validation will be carried out.}

\item{hic}{A scalar >0. The smoothing bandwidth for calculate the loss.
Default is \code{NULL}. If \code{NULL}, then the cross validation will be
carried out.}

\item{detect.parallel}{A logic variable to determine whether the
informative set detection is parallel for sources, default is
\code{FALSE}.}

\item{compute.parallel}{A logic variable to determine whether the Step 1
in our proposed method run with parallel computation, default is
\code{FALSE}.}

\item{ncore}{The integer, the number of cores used for parallel computation.}

\item{seed}{An integer for random seed for data split in infotmative set
detection procedure, default is 111.}

\item{lowc}{Constant used for calculating lower bound of bandwidth according
to our paper, default is 0.05.}

\item{upc}{Constant used for calculating upper bound of bandwidth according
to our paper, default is 1.}

\item{thresh}{The threshold used to determine the sparsity of initial,
default is 1e-10.}

\item{length.out}{The number of grid search points for searching the best
bandwidth, default is 25.}

\item{max.it}{The maximum iteration number for lasso problem via glmnet,
default is 1e6.}

\item{max.round}{The maximum round of expanding the range of bandwidth
selection before warning if provided is not suitable. Default is 5.}

\item{verbose}{A logic variable to control the print of loss during
informative set detection, default is FALSE.}
}
\description{
Internal tlqr parameters
}
