% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{TransferPoolQR}
\alias{TransferPoolQR}
\title{Oracle Trans-pooling QR with a given informative sources set}
\usage{
TransferPoolQR(x.target, y.target, x.source, y.source, u)
}
\arguments{
\item{x.target}{A n*p matrix as the covariate from the target population.}

\item{y.target}{A vector with length n as the response from the target
population.}

\item{x.source}{A list object that each element is a n*p matrix as the
covariate from informative sources.}

\item{y.source}{A list object that each element is a vector with length n
as the covariate from informative sources.}

\item{u}{A scalar in (0,1). The quantile level for all models.}
}
\value{
A vector. The corresponding transfer learning estimator based on
the given auxiliary sources based on Huang 2022.
}
\description{
Oracle Trans-pooling QR with a given informative sources set
}
