% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tune_lasso}
\alias{tune_lasso}
\title{Build the function for tunning the parameter based on given data for lasso
problem.}
\usage{
tune_lasso(x, y, nbar, nfolds = 10, seed = 111)
}
\arguments{
\item{x}{A n*p matrix as the training and validation covariate.}

\item{y}{A vector with length n as the response from training and
validation.}

\item{nbar}{The maximum sample size among all datasets}

\item{nfolds}{The number of folds for the cross validation.}

\item{seed}{An integer. The default is 111. The random seed.}
}
\value{
A scalar. The best lambda value for penalized model.
}
\description{
Build the function for tunning the parameter based on given data for lasso
problem.
}
