% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tune_param}
\alias{tune_param}
\title{Build the function for tunning the parameter based on given data (combine of train+valid)}
\usage{
tune_param(
  x,
  y,
  u,
  size = 0.05,
  lower = -2,
  upper = 2,
  step = 0.05,
  machines = 10,
  seed = 111
)
}
\arguments{
\item{x}{A n*p matrix as the training and validation covariate.}

\item{y}{A vector with length n as the response from training and validation.}

\item{u}{A scalar. The quantile level for model.}

\item{size}{A scalar in [0,1]. default is 0.05, the proportion of validation sample.}

\item{lower}{A scalar. The default is -2, the lower bound for grid search, established by exp(lower).}

\item{upper}{A scalar larget than lower. The default is +2, the upper bound for grid search, established by exp(upper).}

\item{step}{A scalar. The default is 0.05, the length of step for the grid search.}

\item{machines}{An integer. The default is 10, the number of machines for parellel computation.}

\item{seed}{An integer. The default is 111. The random seed.}
}
\value{
A scalar. The best lambda value for penalized model.
}
\description{
Build the function for tunning the parameter based on given data (combine of train+valid)
}
