% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{B737data}
\alias{B737data}
\title{QAR data from B737}
\format{
A data table with 1218 rows and 103 variables:
}
\usage{
B737data
}
\description{
QAR data from Boeing 737 flights that contains the maximum vertical acceleration during landing as the response and other measurements as covariates. Detailed explanation of feature code are in the supplementary material of the paper.
}
\keyword{datasets}
