% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{data_generation}
\alias{data_generation}
\title{Data generation function to implement the simulation settings in our paper}
\usage{
data_generation(
  p,
  n,
  d,
  An,
  M,
  eta,
  cov_type = "auto",
  res_type = "normal",
  seed = 111
)
}
\arguments{
\item{p}{Input scalar. The dimensionality for the covariate.}

\item{n}{Input scalar. The sample size for each datasets, including all the target and sources.}

\item{d}{Input scalar. Control the scale of the Laplace distribution, larger d leads to smaller similarity between the target and informative sets. Details are in the section for simulation.}

\item{An}{Input scalar. The number of informative set.}

\item{M}{Input scalar. The total number of sources.}

\item{eta}{Input scalar. The signal-to-noise ratio for residual.}

\item{cov_type}{Input string. The default is 'auto'. Control the covariance matrix of datasets. The value 'auto' leads to a homogeneous design with auto-covariance; the value 'toeplitz' leads to heterogeneous designs with Toeplitz covariance matrix. No other value is permitted.}

\item{res_type}{Input string. The default is 'normal'. Control the type of distribution for residual. Only one from 'normal' and 'cauchy' is allowed.}

\item{seed}{Input scalar. The default is 111.  The random seed.}
}
\value{
NULL. But will generate correspoding pairs of (X,y) in global enviroment.
}
\description{
Data generation function to implement the simulation settings in our paper
}
