% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{info_detect}
\alias{info_detect}
\title{Informative sources detection}
\usage{
info_detect(
  x_target,
  y_target,
  x_aux_bd,
  y_aux_bd,
  u_target,
  u_aux_bd,
  mode = "real",
  psd = FALSE,
  info_num = NULL,
  epsilon = NULL
)
}
\arguments{
\item{x_target}{A n*p matrix as the covariate from the target population.}

\item{y_target}{A vector with length n as the response from the target population.}

\item{x_aux_bd}{A list object that each element is a n*p matrix as the covariate from informative sources.}

\item{y_aux_bd}{A list object that each element is a vector with length n as the covariate from informative sources.}

\item{u_target}{A scalar in (0,1). The quantile level for model of target population.}

\item{u_aux_bd}{A vector that contains the quantile level for informative populations.}

\item{mode}{A string. Default is 'real'. Only 'simulation' or 'real' is allowed. The difference of mode leads to different operations for the selection of strength of penalty in single-source modeling. 'real' mode takes longer time.}

\item{psd}{A boolean variable. Default is true. Whether the procedure is pseudo.}

\item{info_num}{A integer. Default is NULL. The given number of informative sources under pseudo running.}

\item{epsilon}{A scalar. Default is NULL. The strict level for informative sources detection, larger the value is, less strict the procedure is.}
}
\value{
A vector of index. The index of the informative sources.
}
\description{
The main function for informative sources detection among all datasets. This is a general function that can implement both Pseudo and normal informative sources detection, it depends on the input variables explained following.
}
