% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Q_loss}
\alias{Q_loss}
\title{Build the auxiliary function to compute quantile loss for informative
sources detection}
\usage{
Q_loss(betahat, betaic, X_measure, y_measure, u.target, control = list())
}
\arguments{
\item{betahat}{Estimator for target only from training part.}

\item{betaic}{Estimator for target only from testing part.}

\item{X_measure}{A n*p matrix for covariate of population that is waiting
for the test.}

\item{y_measure}{A vector as the response from the population that is
waiting for the test.}

\item{u.target}{A scalar in (0,1). The quantile level for informative
source detection testing.}

\item{control}{The list object for control, see \code{tlqr.control}.}
}
\value{
A scalar. The quantile loss for given estimator on the given
testing samples.
}
\description{
Build the auxiliary function to compute quantile loss for informative
sources detection
}
