% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rq.fusion}
\alias{rq.fusion}
\title{Fusion model}
\usage{
rq.fusion(
  x_target,
  y_target,
  x_aux_bd,
  y_aux_bd,
  u_target,
  u_aux_bd,
  mode = "real"
)
}
\arguments{
\item{x_target}{A n*p matrix as the covariate from the target population.}

\item{y_target}{A vector with length n as the response from the target population.}

\item{x_aux_bd}{A list object that each element is a n*p matrix as the covariate from informative sources.}

\item{y_aux_bd}{A list object that each element is a vector with length n as the covariate from informative sources.}

\item{u_target}{A scalar in (0,1). The quantile level for model of target population.}

\item{u_aux_bd}{A vector that contains the quantile level for informative populations.}

\item{mode}{A string. Default is 'real'. Only 'simulation' or 'real' is allowed. The difference of mode leads to different operations for the selection of strength of penalty in single-source modeling. 'real' mode takes longer time.}
}
\value{
A vector. The fusion estimator without debias procedure based on the given auxiliary sources.
}
\description{
Function for building fusion model between provided target and sources without the final debias procedure.
}
