% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rq.transfer}
\alias{rq.transfer}
\title{Oracle Trans-Lasso QR with a given informative sources set}
\usage{
rq.transfer(
  x_target,
  y_target,
  x_aux_bd,
  y_aux_bd,
  u_target,
  u_aux_bd,
  h = NULL,
  lambda1 = NULL,
  lambda2 = NULL,
  parallel = FALSE,
  ncore = 20,
  seed = 111
)
}
\arguments{
\item{x_target}{A n*p matrix as the covariate from the target population.}

\item{y_target}{A vector with length n as the response from the target
population.}

\item{x_aux_bd}{A list object that each element is a n*p matrix as the
covariate from informative sources.}

\item{y_aux_bd}{A list object that each element is a vector with length n
as the covariate from informative sources.}

\item{u_target}{A scalar in (0,1). The quantile level for model of target
population.}

\item{u_aux_bd}{A vector that contains the quantile level for informative
populations.}

\item{h}{A vector with length (K+1), where K is the number of auxiliary
datasets. It pre-determine the bandwidth. The default is NULL, if NULL,
cross validation will be carried out.}

\item{lambda1}{A scalar for the penalty strength in fusion step. Default
is null, if null, a cross validation will be carried for searching.}

\item{lambda2}{A scalar for the penalty strength in debiasing step. Default
is null, if null, lambda2 will be calculated based on lambda1 according to
the proportion of sample sizes as described in our paper.}

\item{parallel}{A logic variable, default is FALSE.}

\item{ncore}{The integer, the number of cores used for parallel computation.}

\item{seed}{An integer for random seed, default is 111.}
}
\value{
A vector. The corresponding transfer learning estimator based on
the given auxiliary sources.
}
\description{
Oracle Trans-Lasso QR with a given informative sources set
}
