% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{smoothy}
\alias{smoothy}
\title{Get the smoothened response (from the quantile model to least square model) variable by iterations}
\usage{
smoothy(x, y, u, maxit = 100, reltol = 0.01)
}
\arguments{
\item{x}{A n*p matrix as the covariate for response smoothening.}

\item{y}{A vector with length n as the response that needs to be smoothened.}

\item{u}{A scalar. The quantile level.}

\item{maxit}{A integer. The maximum iteration number.}

\item{reltol}{A scalar. Default is 1e-2. The relative tolerance for the stopping criterion.}
}
\value{
y: A vector as the smoothed response. h: A scalar as the selected bandwidth.
}
\description{
Get the smoothened response (from the quantile model to least square model) variable by iterations
}
